+++ packages/env/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @neondatabase/env
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @neondatabase/config@0.2.1
+
 ## 0.1.1
 
 ### Patch Changes